passphrase modal refactor
