+++ packages/suite/src/components/suite/modals/Passphrase/index.tsx
@@ -1,11 +1,9 @@
-import React, { useState, createRef, useEffect } from 'react';
+import React from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import { useKeyPress } from '@suite-utils/dom';
 import styled, { css } from 'styled-components';
-import { Button, H2, colors, variables, Input, Checkbox } from '@trezor/components-v2';
+import { Button, H2, colors, variables } from '@trezor/components-v2';
 // import { Translation } from '@suite-components/Translation';
-import Loading from '@suite-components/Loading';
 import * as modalActions from '@suite-actions/modalActions';
 import * as discoveryActions from '@wallet-actions/discoveryActions';
 import * as deviceUtils from '@suite-utils/device';
@@ -14,6 +12,7 @@ import { AppState, Dispatch, TrezorDevice } from '@suite-types';
 import Link from '../../Link';
 import ModalWrapper from '@suite-components/ModalWrapper';
 import { PASSPHRASE_URL } from '@suite-constants/urls';
+import PassphraseTypeCard from './components/PassphraseTypeCard';
 
 const Wrapper = styled.div<{ authConfirmation?: boolean }>`
     display: flex;
@@ -41,46 +40,6 @@ const Description = styled.div`
     margin-bottom: 4px;
 `;
 
-const WalletTitle = styled.div`
-    font-size: ${variables.FONT_SIZE.NORMAL};
-    text-align: center;
-    color: ${colors.BLACK0};
-    margin-bottom: 12px;
-`;
-
-const AuthConfirmWrapper = styled.div<{ secondary?: boolean }>`
-    display: flex;
-    flex: 1;
-    max-width: 360px;
-    flex-direction: column;
-    padding: 32px 24px;
-    align-items: center;
-    border-radius: 6px;
-    background: ${colors.BLACK96};
-`;
-
-const Col = styled.div<{ secondary?: boolean }>`
-    display: flex;
-    flex: 1;
-    width: 320px;
-    flex-direction: column;
-    padding: 32px 24px;
-    align-items: center;
-    border: solid 2px ${colors.BLACK96};
-    border-radius: 6px;
-    justify-self: center;
-
-    ${props =>
-        props.secondary &&
-        css`
-            background: ${colors.BLACK96};
-        `}
-        
-    @media screen and (max-width: ${variables.SCREEN_SIZE.MD}) {
-        width: 100%;
-    }
-`;
-
 const Content = styled.div`
     display: flex;
     flex: 1;
@@ -89,21 +48,61 @@ const Content = styled.div`
     margin-bottom: 32px;
 `;
 
-const InputWrapper = styled(Content)`
-    margin-bottom: 32px;
-    width: 100%;
-`;
+// const WalletTitle = styled.div`
+//     font-size: ${variables.FONT_SIZE.NORMAL};
+//     text-align: center;
+//     color: ${colors.BLACK0};
+//     margin-bottom: 12px;
+// `;
 
-const Actions = styled.div`
-    width: 100%;
-    /* margin-top: 20px; */
-`;
+// const AuthConfirmWrapper = styled.div<{ secondary?: boolean }>`
+//     display: flex;
+//     flex: 1;
+//     max-width: 360px;
+//     flex-direction: column;
+//     padding: 32px 24px;
+//     align-items: center;
+//     border-radius: 6px;
+//     background: ${colors.BLACK96};
+// `;
 
-const Padding = styled.div`
-    display: flex;
-    flex-direction: column;
-    margin: 0px 24px;
-`;
+// const Col = styled.div<{ secondary?: boolean }>`
+//     display: flex;
+//     flex: 1;
+//     width: 320px;
+//     flex-direction: column;
+//     padding: 32px 24px;
+//     align-items: center;
+//     border: solid 2px ${colors.BLACK96};
+//     border-radius: 6px;
+//     justify-self: center;
+
+//     ${props =>
+//         props.secondary &&
+//         css`
+//             background: ${colors.BLACK96};
+//         `}
+
+//     @media screen and (max-width: ${variables.SCREEN_SIZE.MD}) {
+//         width: 100%;
+//     }
+// `;
+
+// const InputWrapper = styled(Content)`
+//     margin-bottom: 32px;
+//     width: 100%;
+// `;
+
+// const Actions = styled.div`
+//     width: 100%;
+//     /* margin-top: 20px; */
+// `;
+
+// const Padding = styled.div`
+//     display: flex;
+//     flex-direction: column;
+//     margin: 0px 24px;
+// `;
 
 const mapStateToProps = (state: AppState) => ({
     devices: state.devices,
@@ -133,96 +132,93 @@ const Passphrase = (props: Props) => {
         device.features.capabilities &&
         device.features.capabilities.includes('Capability_PassphraseEntry');
 
-    const [submitted, setSubmitted] = useState(false);
-    const [enabled, setEnabled] = useState(!authConfirmation);
-    const [value, setValue] = useState('');
-    const [showPassword, setShowPassword] = useState(false);
-    const inputType = showPassword ? 'text' : 'password';
-    const enterPressed = useKeyPress('Enter');
-    const ref = createRef<HTMLInputElement>();
-
-    useEffect(() => {
-        if (ref && ref.current) {
-            ref.current.focus();
-        }
-    }, [ref]);
-
-    if (submitted) {
-        return <Loading />;
-    }
-
-    const submit = (passphraseOnDevice?: boolean) => {
-        if (!enabled) return;
-        setSubmitted(true);
-        props.onPassphraseSubmit(value, passphraseOnDevice);
-    };
-
-    if (enterPressed) {
-        submit();
-    }
+    // const [submitted, setSubmitted] = useState(false);
+    // const [enabled, setEnabled] = useState(!authConfirmation);
+    // const [value, setValue] = useState('');
+    // const [showPassword, setShowPassword] = useState(false);
+    // const inputType = showPassword ? 'text' : 'password';
+    // const enterPressed = useKeyPress('Enter');
+    // const ref = createRef<HTMLInputElement>();
 
     // TODO: translations
-    const INPUT_PLACEHOLDER = 'Enter passphrase';
+    // const INPUT_PLACEHOLDER = 'Enter passphrase';
 
     if (authConfirmation || stateConfirmation) {
         // show borderless one-column modal for confirming passphrase and state confirmation
         return (
-            <AuthConfirmWrapper>
-                <WalletTitle>
-                    {stateConfirmation ? 'Enter passphrase' : 'Confirm empty hidden wallet'}
-                </WalletTitle>
-                <Content>
-                    {stateConfirmation
+            <PassphraseTypeCard
+                authConfirmation={authConfirmation}
+                title={stateConfirmation ? 'Enter passphrase' : 'Confirm empty hidden wallet'}
+                description={
+                    stateConfirmation
                         ? 'Unlock'
-                        : 'This hidden Wallet is empty. To make sure you are in the correct Wallet, confirm Passphrase.'}
-                </Content>
-                <Padding>
-                    {authConfirmation && (
-                        <Content>
-                            <Checkbox onClick={() => setEnabled(!enabled)} isChecked={enabled}>
-                                I understand passphrase is not saved anywhere and can’t be restored.
-                            </Checkbox>
-                        </Content>
-                    )}
-                    <InputWrapper>
-                        <Input
-                            onChange={event => setValue(event.target.value)}
-                            placeholder={INPUT_PLACEHOLDER}
-                            type={inputType}
-                            value={value}
-                            innerRef={ref}
-                            display="block"
-                            variant="small"
-                            button={{
-                                iconSize: 18,
-                                icon: showPassword ? 'HIDE' : 'SHOW',
-                                onClick: () => setShowPassword(!showPassword),
-                            }}
-                        />
-                    </InputWrapper>
-                    <Actions>
-                        <Button
-                            isDisabled={!enabled}
-                            variant="primary"
-                            onClick={() => submit()}
-                            fullWidth
-                        >
-                            Confirm passphrase
-                        </Button>
-                        {onDeviceOffer && (
-                            <Button
-                                isDisabled={!enabled}
-                                variant="secondary"
-                                onClick={() => submit(true)}
-                                fullWidth
-                            >
-                                Enter passphrase on device
-                            </Button>
-                        )}
-                    </Actions>
-                </Padding>
-            </AuthConfirmWrapper>
+                        : 'This hidden Wallet is empty. To make sure you are in the correct Wallet, confirm Passphrase.'
+                }
+                submitLabel="Confirm passphrase"
+                colorVariant="secondary"
+                offerPassphraseOnDevice={onDeviceOffer}
+                onPassphraseSubmit={props.onPassphraseSubmit}
+                singleColModal
+                showPassphraseInput
+            />
         );
+        // return (
+        //     <AuthConfirmWrapper>
+        //         <WalletTitle>
+        //             {stateConfirmation ? 'Enter passphrase' : 'Confirm empty hidden wallet'}
+        //         </WalletTitle>
+        //         <Content>
+        //             {stateConfirmation
+        //                 ? 'Unlock'
+        //                 : 'This hidden Wallet is empty. To make sure you are in the correct Wallet, confirm Passphrase.'}
+        //         </Content>
+        //         <Padding>
+        //             {authConfirmation && (
+        //                 <Content>
+        //                     <Checkbox onClick={() => setEnabled(!enabled)} isChecked={enabled}>
+        //                         I understand passphrase is not saved anywhere and can’t be restored.
+        //                     </Checkbox>
+        //                 </Content>
+        //             )}
+        //             <InputWrapper>
+        //                 <Input
+        //                     onChange={event => setValue(event.target.value)}
+        //                     placeholder={INPUT_PLACEHOLDER}
+        //                     type={inputType}
+        //                     value={value}
+        //                     innerRef={ref}
+        //                     display="block"
+        //                     variant="small"
+        //                     button={{
+        //                         iconSize: 18,
+        //                         icon: showPassword ? 'HIDE' : 'SHOW',
+        //                         onClick: () => setShowPassword(!showPassword),
+        //                     }}
+        //                 />
+        //             </InputWrapper>
+        //             <Actions>
+        //                 <Button
+        //                     isDisabled={!enabled}
+        //                     variant="primary"
+        //                     onClick={() => submit()}
+        //                     fullWidth
+        //                 >
+        //                     Confirm passphrase
+        //                 </Button>
+        //                 {onDeviceOffer && (
+        //                     <Button
+        //                         isDisabled={!enabled}
+        //                         variant="secondary"
+        //                         onClick={() => submit(true)}
+        //                         fullWidth
+        //                     >
+        //                         Enter passphrase on device
+        //                     </Button>
+        //                 )}
+        //             </Actions>
+        //         </Padding>
+        //     </AuthConfirmWrapper>
+        // );
     }
 
     // show 2-column modal for selecting between standard and hidden wallets
@@ -247,19 +243,36 @@ const Passphrase = (props: Props) => {
                 </Link>
                 <WalletsWrapper>
                     {noPassphraseOffer && (
-                        <Col>
-                            <WalletTitle>No-passphrase Wallet</WalletTitle>
-                            <Content>
-                                To access standard (no-passphrase) Wallet click the button below.
-                            </Content>
-                            <Actions>
-                                <Button variant="primary" fullWidth onClick={() => submit()}>
-                                    Access standard Wallet
-                                </Button>
-                            </Actions>
-                        </Col>
+                        <PassphraseTypeCard
+                            title="No-passphrase Wallet"
+                            description="To access standard (no-passphrase) Wallet click the button below."
+                            submitLabel="Access standard Wallet"
+                            colorVariant="primary"
+                            onPassphraseSubmit={props.onPassphraseSubmit}
+                        />
+                        // <Col>
+                        //     <WalletTitle>No-passphrase Wallet</WalletTitle>
+                        //     <Content>
+                        //         To access standard (no-passphrase) Wallet click the button below.
+                        //     </Content>
+                        //     <Actions>
+                        //         <Button variant="primary" fullWidth onClick={() => submit()}>
+                        //             Access standard Wallet
+                        //         </Button>
+                        //     </Actions>
+                        // </Col>
                     )}
-                    <Col secondary>
+                    <PassphraseTypeCard
+                        title="Passphrase (hidden) wallet"
+                        description="Enter existing passphrase to access existing hidden Wallet. Or enter new
+                            passphrase to create a new hidden Wallet."
+                        submitLabel="Access Hidden Wallet"
+                        colorVariant="secondary"
+                        showPassphraseInput
+                        onPassphraseSubmit={props.onPassphraseSubmit}
+                    />
+
+                    {/* <Col secondary>
                         <WalletTitle>Passphrase (hidden) wallet</WalletTitle>
                         <Content>
                             Enter existing passphrase to access existing hidden Wallet. Or enter new
@@ -301,7 +314,7 @@ const Passphrase = (props: Props) => {
                                 </Button>
                             )}
                         </Actions>
-                    </Col>
+                    </Col> */}
                 </WalletsWrapper>
             </Wrapper>
         </ModalWrapper>
